+++ src/shared/particles/base_particles.hpp
@@ -68,11 +68,11 @@ namespace SPH
         }
     }
     //=================================================================================================//
-    template <typename GlobalVariableType>
-    void BaseParticles::
-        registerGlobalVariable(const std::string &variable_name, GlobalVariableType initial_value)
+    template <typename DataType>
+    DataType *BaseParticles::
+        registerGlobalVariable(const std::string &variable_name, DataType initial_value)
     {
-        constexpr int type_index = DataTypeIndex<GlobalVariableType>::value;
+        constexpr int type_index = DataTypeIndex<DataType>::value;
         bool isRegistered = false;
 
         for (auto& variable : std::get<type_index>(all_global_data_))
@@ -80,24 +80,25 @@ namespace SPH
             if (variable->getName() == variable_name)
             {
                 isRegistered = true;
-                std::cout << "\n Error: the variable '" << variable_name << "' has already been registered!" << std::endl;
-                std::cout << __FILE__ << ':' << __LINE__ << std::endl;
-                exit(1);
+                return variable->getValue();
             }
         }
 
         if (!isRegistered)
         {
-            UniquePtrsKeeper<GlobalVariable<GlobalVariableType>>& container = std::get<type_index>(all_global_data_ptr_);
-            GlobalVariable<GlobalVariableType> *contained_data = container.template createPtr<GlobalVariable<GlobalVariableType>>(variable_name, initial_value);
+            UniquePtrsKeeper<GlobalVariable<DataType>>& container = std::get<type_index>(all_global_data_ptr_);
+            GlobalVariable<DataType> *contained_data = container.template createPtr<GlobalVariable<DataType>>(variable_name, initial_value);
             std::get<type_index>(all_global_data_).push_back(contained_data);
+            return contained_data->getValue();
         }
+
+        return nullptr;
     }
      //=================================================================================================//
-    template <typename GlobalVariableType>
-    GlobalVariableType *BaseParticles::getGlobalVariableByName(const std::string &variable_name)
+    template <typename DataType>
+    DataType *BaseParticles::getGlobalVariableByName(const std::string &variable_name)
     {
-        constexpr int type_index = DataTypeIndex<GlobalVariableType>::value;
+        constexpr int type_index = DataTypeIndex<DataType>::value;
         bool isRegistered = false;
 
         for (auto& variable : std::get<type_index>(all_global_data_))
@@ -116,6 +117,8 @@ namespace SPH
             std::cout << __FILE__ << ':' << __LINE__ << std::endl;
             return nullptr;
         }
+
+        return nullptr;
     }
     //=================================================================================================//
     template <typename VariableType>
