+++ tests/user_examples/test_2d_diffusion_RobinBC/test_2d_diffusion_RobinBC.h
@@ -1,15 +1,12 @@
 /**
- * @file 	2d_diffusion_test_with_RobinBC.h
- * @brief 	This is the head files used by 2d_diffusion_test_with_RobinBC.cpp.
+ * @file 	test_2d_diffusion_RobinBC.h
+ * @brief 	This is the head files used by test_2d_diffusion_RobinBC.cpp.
  * @author	Chenxi Zhao, Bo Zhang, Chi Zhang and Xiangyu Hu
  */
 #ifndef	DIFFUSION_TEST_WITH_ROBINBC_H
 #define DIFFUSION_TEST_WITH_ROBINBC_H
 
 #include "sphinxsys.h"
-#include "particle_dynamics_diffusion_reaction_with_boundary.h"
-#include "particle_dynamics_diffusion_reaction_with_boundary.hpp"
-
 using namespace SPH;
 
 //----------------------------------------------------------------------
@@ -160,14 +157,18 @@ public:
 };
 
 class RobinWallBoundaryInitialCondition
-	: public DiffusionReactionInitialConditionWithRobin<WallParticles>
+	: public DiffusionReactionInitialCondition<WallParticles>
 {
 protected:
 	size_t phi_;
+	StdLargeVec<Real>& convection_;
+	Real& T_infinity_;
 
 public:
 	RobinWallBoundaryInitialCondition(SolidBody& diffusion_body) :
-		DiffusionReactionInitialConditionWithRobin<WallParticles>(diffusion_body)
+		DiffusionReactionInitialCondition<WallParticles>(diffusion_body),
+		convection_(*(this->particles_->template getVariableByName<Real>("Convection"))),
+		T_infinity_(*(this->particles_->template getGlobalVariableByName<Real>("T_infinity")))
 	{
 		phi_ = particles_->diffusion_reaction_material_.AllSpeciesIndexMap()["Phi"];
 	}
@@ -179,7 +180,7 @@ public:
 		if (pos_[index_i][1] < 0 && pos_[index_i][0] > 0.45 * L && pos_[index_i][0] < 0.55 * L)
 		{
 			convection_[index_i] = convection;
-			T_infinity_.setValue(T_infinity);
+			T_infinity_ = T_infinity;
 		}
 	}
 };
@@ -194,8 +195,11 @@ class DiffusionBodyRelaxation
 	: public DiffusionRelaxationRK2<ComplexInteraction<SolidDiffusionInner, SolidDiffusionDirichlet, SolidDiffusionRobin>>
 {
 public:
-	explicit DiffusionBodyRelaxation(BaseInnerRelation& inner_relation, BaseContactRelation& body_contact_relation_Dirichlet, BaseContactRelation& body_contact_relation_Robin)
-		: DiffusionRelaxationRK2<ComplexInteraction<SolidDiffusionInner, SolidDiffusionDirichlet, SolidDiffusionRobin>>(inner_relation, body_contact_relation_Dirichlet, body_contact_relation_Robin) {};
+	explicit DiffusionBodyRelaxation(BaseInnerRelation& inner_relation,
+									 BaseContactRelation& body_contact_relation_Dirichlet,
+									 BaseContactRelation& body_contact_relation_Robin)
+		: DiffusionRelaxationRK2<ComplexInteraction<SolidDiffusionInner, SolidDiffusionDirichlet, SolidDiffusionRobin>>(
+			inner_relation, body_contact_relation_Dirichlet, body_contact_relation_Robin) {};
 	virtual ~DiffusionBodyRelaxation() {};
 };
 //----------------------------------------------------------------------
@@ -214,7 +218,8 @@ public:
 		for (size_t i = 0; i < number_of_observation_points; ++i)
 		{
 			Vec2d point_coordinate(0.5 * L, range_of_measure * Real(i) /
-				Real(number_of_observation_points - 1) + start_of_measure);
+                                                    Real(number_of_observation_points - 1) +
+                                                start_of_measure);
 			positions_.push_back(point_coordinate);
 		}
 	}
